refactored fiels
