+++ src/setup.py
@@ -2,16 +2,29 @@ from typing import List
 
 from common.color_type_enum import ColourType
 
+# setup_sides: List[List[ColourType]] = [
+#     [ColourType.yellow, ColourType.white, ColourType.orange, ColourType.white, ColourType.white, ColourType.white,
+#      ColourType.orange, ColourType.white, ColourType.yellow],
+#     [ColourType.blue, ColourType.yellow, ColourType.white, ColourType.red, ColourType.red, ColourType.red,
+#      ColourType.green, ColourType.yellow, ColourType.white],
+#     [ColourType.green, ColourType.orange, ColourType.blue, ColourType.blue, ColourType.blue, ColourType.green,
+#      ColourType.blue, ColourType.blue, ColourType.yellow],
+#     [ColourType.white, ColourType.orange, ColourType.red, ColourType.yellow, ColourType.orange, ColourType.orange,
+#      ColourType.green, ColourType.red, ColourType.green],
+#     [ColourType.white, ColourType.green, ColourType.red, ColourType.yellow, ColourType.green, ColourType.orange,
+#      ColourType.yellow, ColourType.blue, ColourType.blue],
+#     [ColourType.red, ColourType.green, ColourType.red, ColourType.green, ColourType.yellow, ColourType.blue,
+#      ColourType.orange, ColourType.red, ColourType.orange]]
 setup_sides: List[List[ColourType]] = [
-    [ColourType.white, ColourType.white, ColourType.blue, ColourType.white, ColourType.white, ColourType.white,
-     ColourType.red, ColourType.white, ColourType.orange],
-    [ColourType.red, ColourType.orange, ColourType.yellow, ColourType.blue, ColourType.blue, ColourType.blue,
-     ColourType.green, ColourType.blue, ColourType.green],
-    [ColourType.green, ColourType.green, ColourType.blue, ColourType.blue, ColourType.orange, ColourType.yellow,
-     ColourType.blue, ColourType.orange, ColourType.yellow],
-    [ColourType.white, ColourType.orange, ColourType.orange, ColourType.yellow, ColourType.green, ColourType.green,
-     ColourType.white, ColourType.red, ColourType.blue],
-    [ColourType.white, ColourType.red, ColourType.yellow, ColourType.green, ColourType.red, ColourType.yellow,
-     ColourType.green, ColourType.yellow, ColourType.yellow],
-    [ColourType.orange, ColourType.orange, ColourType.red, ColourType.red, ColourType.yellow, ColourType.green,
-     ColourType.red, ColourType.red, ColourType.orange]]
+    [ColourType.orange, ColourType.white, ColourType.green, ColourType.yellow, ColourType.white, ColourType.yellow,
+     ColourType.orange, ColourType.white, ColourType.green],
+    [ColourType.white, ColourType.orange, ColourType.blue, ColourType.green, ColourType.blue, ColourType.red,
+     ColourType.white, ColourType.red, ColourType.green],
+    [ColourType.blue, ColourType.orange, ColourType.red, ColourType.yellow, ColourType.orange, ColourType.white,
+     ColourType.white, ColourType.red, ColourType.orange],
+    [ColourType.yellow, ColourType.blue, ColourType.blue, ColourType.white, ColourType.green, ColourType.orange,
+     ColourType.yellow, ColourType.orange, ColourType.blue],
+    [ColourType.yellow, ColourType.blue, ColourType.red, ColourType.green, ColourType.red, ColourType.green,
+     ColourType.green, ColourType.blue, ColourType.orange],
+    [ColourType.white, ColourType.blue, ColourType.red, ColourType.yellow, ColourType.yellow, ColourType.green,
+     ColourType.yellow, ColourType.red, ColourType.red]]
